FDG-11374 update tests
