+++ src/components/dtg-table/test-data.js
@@ -444,6 +444,7 @@ export const mockPaginatedTableProps = {
     endpoint: 'v1/accounting/dts/dts_table_1',
     rowCount: 21000,
   },
+  config: {},
   dateRange: dateRange,
   serverSidePagination: 'v1/accounting/dts/dts_table_1',
   shouldPage: true,
@@ -481,6 +482,7 @@ export const mockReactTableProps_rawData = {
     endpoint: 'v1/accounting/dts/dts_table_1',
     rowCount: 4000,
   },
+  config: {},
   dateRange: dateRange,
   serverSidePagination: null,
   shouldPage: true,
@@ -494,6 +496,7 @@ export const mockReactTableProps_rawData_smallTable = {
     endpoint: 'v1/accounting/dts/dts_table_1',
     rowCount: 1000,
   },
+  config: {},
   dateRange: dateRange,
   dePagianted: null,
   serverSidePagination: null,
@@ -508,6 +511,7 @@ export const mockReactTableProps_rawData_pivotTable = {
     endpoint: 'v1/accounting/dts/dts_table_1',
     rowCount: 1000,
   },
+  config: {},
   dateRange: dateRange,
   dePagianted: null,
   serverSidePagination: null,
